requirements.txt
# Core Dependencies
python-telegram-bot==20.7
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1

# Caching & Queue
redis==5.0.1

# HTTP & API
requests==2.31.0
aiohttp==3.9.1
httpx==0.25.2

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Date/Time
python-dateutil==2.8.2
pytz==2023.3

# Text Processing & Formatting
jinja2==3.1.2
markdown==3.5.1

# PDF Generation
reportlab==4.0.7

# Logging
loguru==0.7.2

# Environment & Config
pydantic==2.5.3
pydantic-settings==2.1.0

# Testing (for later)
pytest==7.4.3
pytest-asyncio==0.21.1

# Code Quality
black==23.12.1
flake8==6.1.0
